requirements.txt
xgboost>=2.0.0
shap>=0.44.0
optuna>=3.5.0
mlflow>=2.11.0
scikit-learn>=1.4.0
imbalanced-learn>=0.12.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=15.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
pyyaml>=6.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
joblib>=1.3.0
python-dotenv>=1.0.0
pytest>=8.0.0
